Create custom ComboBox and TextBox for VerseControl
