+++ SIL.Windows.Forms/Widgets/SafeComboBox.cs
@@ -1,54 +1,16 @@
 using System;
 using System.ComponentModel;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
 using SIL.PlatformUtilities;
 
 namespace SIL.Windows.Forms.Widgets
 {
-	public delegate void CopyPasteEmbeddedTextBox();
-
 	/// <summary>
 	/// Replaces a Windows combo box control (System.Windows.Forms.ComboBox)
 	/// Returns an empty string when get_Text generates an IndexOutOfRangeException (FB 6643, 6750)
 	/// </summary>
 	public class SafeComboBox : ComboBox
 	{
-		public event CopyPasteEmbeddedTextBox CopyEvent;
-		public event CopyPasteEmbeddedTextBox PasteEvent;
-
-		[DllImport("user32.dll", EntryPoint = "FindWindowExA", CallingConvention = CallingConvention.StdCall, CharSet = CharSet.Ansi)]
-		private static extern IntPtr FindWindowEx(IntPtr hwndParent, IntPtr hwndChildAfter, string lpszClass, string lpszWindow);
-		private TextBoxHandle textBoxHandle;
-
-		/// <summary>
-		/// Get handle of TextBox that is embedded in the ComboBox,
-		/// and is needed for copying and pasting VerseRefs using context menu
-		/// </summary>
-		private void GetTextBoxHandle()
-		{
-			if (textBoxHandle == null)
-			{
-				IntPtr embeddedTextBox = FindWindowEx(this.Handle, IntPtr.Zero, "EDIT", null);
-
-				textBoxHandle = new TextBoxHandle();
-				textBoxHandle.AssignHandle(embeddedTextBox);
-
-				textBoxHandle.CopyEvent += OnCopyEvent;
-				textBoxHandle.PasteEvent += OnPasteEvent;
-			}
-		}
-
-		protected virtual void OnCopyEvent()
-		{
-			CopyEvent?.Invoke();
-		}
-
-		protected virtual void OnPasteEvent()
-		{
-			PasteEvent?.Invoke();
-		}
-
 		/// <summary>
 		/// Occurs before mouse down.
 		/// </summary>
@@ -56,9 +18,6 @@ namespace SIL.Windows.Forms.Widgets
 
 		protected override void OnMouseDown(MouseEventArgs e)
 		{
-			if (Platform.IsWindows)
-				GetTextBoxHandle();
-
 			if (BeforeMouseDown != null)
 				BeforeMouseDown(this, e);
 
@@ -104,41 +63,6 @@ namespace SIL.Windows.Forms.Widgets
 				{
 				}
 			}
-		}
-	}
-
-	/// <summary>
-	/// Handle of the TextBox that is embedded in the ComboBox
-	/// </summary>
-	public class TextBoxHandle : NativeWindow
-	{
-		public event CopyPasteEmbeddedTextBox CopyEvent;
-		public event CopyPasteEmbeddedTextBox PasteEvent;
-
-		protected override void WndProc(ref Message m)
-		{
-			switch (m.Msg)
-			{
-				case (0x301): //WM_COPY
-					OnCopyEvent();
-					break;
-				case (0x302): //WM_PASTE
-					OnPasteEvent();
-					break;
-				default:
-					base.WndProc(ref m);
-					break;
-			}
-		}
-
-		protected virtual void OnCopyEvent()
-		{
-			CopyEvent?.Invoke();
-		}
-
-		protected virtual void OnPasteEvent()
-		{
-			PasteEvent?.Invoke();
 		}
 	}
 }
